Multitenancy

+++ src/main/java/com/example/aichatbot/dto/ChatResponseDto.java
@@ -3,7 +3,7 @@ package com.example.aichatbot.dto;
 /**
  * Response DTO for chat messages and general API responses
  */
-public record ChatResponseDto(String response, Integer conversationId) {
+public record ChatResponseDto(String response, Long conversationId) {
     public ChatResponseDto(String response) {
         this(response, null);
     }
